remove the iconbutton
